% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_nhanes.R
\name{search_nhanes}
\alias{search_nhanes}
\title{Search NHANES Data Using Your Search Query}
\usage{
search_nhanes(query)
}
\arguments{
\item{query}{Any search term in text that the user wants information. There 
should be one search query at a time e.g "demographic", "weight", 
"spirometry". To get a list of all data available in a given publicly 
available dataset, user can search for the dataset name e.g. "demographics",
or "examination", or "laboratory" or "questionnaire".}
}
\value{
A data frame listing matching NHANES files and associated metadata. 
If no matches are found, user will get a message that "No matches found for:".
}
\description{
Generates a list of all the NHANES datasets that match the your search term.
}
\details{
This helps users to inspect what datasets are present in the NHANES data 
related to their search term.They can then use the result to identify the 
dataset that they want. Then using the search result users can pick the year,
dataset, component, or dataset_id of that particular data they want. These 
inputs will go into the preview function and the data download function.


This function uses the internal NHANES mapping table to resolve
dataset availability across survey cycles. It does not perform any
downloads or external requests.
}
\examples{
# Example 1. Search for term demographic
search_nhanes("demographic")
 
# Example 2. Search for term spirometry
search_nhanes("spirometry")
\donttest{
# Example 3. Search for term tuberculosis
search_nhanes("tuberculosis")

# Example 4. Search for term blood_pressure
search_nhanes("blood_pressure")

# Example 5. Search for the whole list within the laboratory dataset
search_nhanes("laboratory") 
}
}
